+++ doc/rest-api.md
@@ -153,7 +153,7 @@ by the object itself.
 To filter your results on certain values, filter is implemented for collections.
 A `filter` argument can be passed to a GET query against a collection.
 
-FIltering is available for the instance and image endpoints.
+Filtering is available for the instance and image endpoints.
 
 There is no default value for filter which means that all results found will
 be returned. The following is the language used for the filter argument:
